refactor: namings, cyphal cleaning

+++ Src/drivers/rcpwm/rcpwm.cpp
@@ -55,8 +55,7 @@ int16_t Driver::RCPWM::get_current_angle(uint16_t max_servo_angle, uint8_t pin_i
                                 channels[pin_idx].min,
                                 channels[pin_idx].max,
                                 -1.0f, +1.0f);
-
-    float angle_deg = normalized * (max_servo_angle / 2.0f);
+    float angle_deg = normalized * (static_cast<float>(max_servo_angle) / 2.0f);
     return static_cast<int16_t>(angle_deg);
 }
